doc: Update dependency installation for Debian/Ubuntu
According to the description for pkg-config, "pkgconf is a
replacement for pkg-config, providing additional functionality
while also maintaining compatibility. This package only provides
a dependency link to the pkgconf package to help with package
upgrades. It can be safely removed."

Thus doc/build-unix.md and depends/README.md are updated.